CHROMIUM: chromeos_acpi: Don't evaluate the "\\S3UE" object
Per firmware team, the S3UE was used on some old boards (cyan/link/rambi)
to set an nvs value, which was then used to decide if they had to tell
the EC to set a specific usb charge mode e.g.
https://chromium.googlesource.com/chromiumos/third_party/coreboot/+/chromeos-2016.05/src/mainboard/google/cyan/smihandler.c#100

But, that is not required on any of the recent boards and thus the BIOS
does not have this method (and it won't have in future). As a result,
on all recent boards, we see a "failed to enable S3 USB wake" message in
kernel bootup logs. Lets get rid of it on newer kernels.

BUG=b:115755982
TEST=Test that no "failed to enable S3 USB wake" message on hatch.

Change-Id: I2291c4370dba1d312214858b6774aec506a81197
Signed-off-by: Rajat Jain <rajatja@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1436556
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>