CHROMIUM: mm: protect VMA modifications using VMA sequence count
The VMA sequence count has been introduced to allow fast detection of
VMA modification when running a page fault handler without holding
the mmap_sem.

This patch provides protection against the VMA modification done in :
	- madvise()
	- mpol_rebind_policy()
	- vma_replace_policy()
	- change_prot_numa()
	- mlock(), munlock()
	- mprotect()
	- mmap_region()
	- collapse_huge_page()
	- userfaultd registering services

In addition, VMA fields which will be read during the speculative fault
path needs to be written using WRITE_ONCE to prevent write to be split
and intermediate values to be pushed to other CPUs.

BUG=b:179777419
TEST=build 5.4 for eve

[SR: fixed conflicts for 5.4]
 Conflicts:
	fs/userfaultfd.c
	mm/madvise.c
Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: Idcb364c2ec3cacf4de1d681748a0ed3317997b51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688236
Reviewed-by: Joel Fernandes <joelaf@google.com>