CHROMIUM: mm: introduce __vm_normal_page()
When dealing with the speculative fault path we should use the VMA's field
cached value stored in the vm_fault structure.

Currently vm_normal_page() is using the pointer to the VMA to fetch the
vm_flags value. This patch provides a new __vm_normal_page() which is
receiving the vm_flags flags value as parameter.

Note: The speculative path is turned on for architecture providing support
for special PTE flag. So only the first block of vm_normal_page is used
during the speculative path.

BUG=b:179777419
TEST=build 5.4 for eve

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: 5.4: fix conflict against 25b2995a35b mm: remove MEMORY_DEVICE_PUBLIC support]
 Conflicts:
	include/linux/mm.h
	mm/memory.c
Change-Id: I8371c27c0a52e9c8a1cc045cd893bdcddbd352fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688242
Reviewed-by: Joel Fernandes <joelaf@google.com>