CHROMIUM: mm: protect the RB tree with a sequence lock
Introducing a per mm_struct seqlock, mm_seq field, to protect the changes
made in the MM RB tree. This allows to walk the RB tree without grabbing
the mmap_sem, and on the walk is done to double check that sequence counter
was stable during the walk.

The mm seqlock is held while inserting and removing entries into the MM RB
tree.  Later in this series, it will be check when looking for a VMA
without holding the mmap_sem.

This is based on the initial work from Peter Zijlstra:
https://lore.kernel.org/linux-mm/20100104182813.479668508@chello.nl/

BUG=b:179777419
TEST=build 5.4 for eve

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: I81149ae7a088d278bbc06e5570385df17270fe05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688245
Reviewed-by: Joel Fernandes <joelaf@google.com>