CHROMIUM: mm: Introduce find_vma_rcu()
This allows to search for a VMA structure without holding the mmap_sem.

The search is repeated while the mm seqlock is changing and until we found
a valid VMA.

While under the RCU protection, a reference is taken on the VMA, so the
caller must call put_vma() once it not more need the VMA structure.

At the time a VMA is inserted in the MM RB tree, in vma_rb_insert(), a
reference is taken to the VMA by calling get_vma().

When removing a VMA from the MM RB tree, the VMA is not release immediately
but at the end of the RCU grace period through vm_rcu_put(). This ensures
that the VMA remains allocated until the end the RCU grace period.

Since the vm_file pointer, if valid, is released in put_vma(), there is no
guarantee that the file pointer will be valid on the returned VMA.

BUG=b:179777419
TEST=build 5.4 for eve

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: trivial context conflict]
 Conflicts:
	mm/mmap.c
Change-Id: I587e8620904f35928d84b80cee69c6ef3c19e40e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688247
Reviewed-by: Joel Fernandes <joelaf@google.com>