CHROMIUM: mm: don't do swap readahead during speculative page fault
Vinayak Menon faced a panic because one thread was page faulting a page in
swap, while another one was mprotecting a part of the VMA leading to a VMA
split.
This raise a panic in swap_vma_readahead() because the VMA's boundaries
were not more matching the faulting address.

To avoid this, if the page is not found in the swap, the speculative page
fault is aborted to retry a regular page fault.

BUG=b:179777419
TEST=build 5.4 for eve

Reported-by: Vinayak Menon <vinmenon@codeaurora.org>
Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: I81c80799a9f2e164a25fb4c32ab7130fbdcc3066
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688249
Reviewed-by: Joel Fernandes <joelaf@google.com>