CHROMIUM: powerpc/mm: add speculative page fault
This patch enable the speculative page fault on the PowerPC
architecture.

This will try a speculative page fault without holding the mmap_sem,
if it returns with VM_FAULT_RETRY, the mmap_sem is acquired and the
traditional page fault processing is done.

The speculative path is only tried for multithreaded process as there is no
risk of contention on the mmap_sem otherwise.

BUG=b:179777419
TEST=build 5.4 for eve

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: Icb3fcee744fc26e2d73f1237838ff71fa3315414
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2688256
Reviewed-by: Joel Fernandes <joelaf@google.com>