UPSTREAM: drm/i915: Fix HSW+ DP MSA YCbCr colorspace indication
Looks like we're currently setting the MSA to xvYCC BT.709 instead
of the YCbCr BT.601 claimed by the comment. But even that comment
is wrong since we configure the CSC matrix to BT.709.

Let's remove the bogus statement from the comment and fix the
MSA to indicate YCbCr BT.709 so that it matches the actual
pixel data we're transmitting.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20190718145053.25808-3-ville.syrjala@linux.intel.com
Reviewed-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
(cherry picked from commit 646d3dc855212855b708da4d481051a5e8cd1946)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:146438784, b:146143855, b:145022885, b:144953588
TEST=Boot to graphics on TGL board provided Mesa Iris enabled see: b:141490430
TEST=Run WebGL benchmark with 1k+ fish should not show blue fish
TEST=Font should be clear when typing on login screen

Change-Id: Id1acc51eceab3588840656721b0471f00bfab193