CHROMIUM: iwl7000: mvm: add support for single antenna diversity
There are products which have a single chain with 2 antennas.
In these products, we need to inform the FW that the device has the
single antenna diversity(SAD) feature. In the future, we will read
the active antenna from a BIOS configuration. Currently, we use a
default configuration which means that the FW decides which antenna to use.

Change-Id: Ida7891d0cc5c4fa3e5abcab6e536ae8b5f89c3ea
Signed-off-by: Haim Dreyfuss <haim.dreyfuss@intel.com>
iwl7000-tree: f52c8ec6a737c110f09743290f8518036bb39b0b