CHROMIUM: iwl7000: mvm: add support for single antenna diversity
There are products which have a single chain with 2 antennas.
In these products, we need to inform the FW that the device has the
single antenna diversity(SAD) feature. In the future, we will read
the active antenna from a BIOS configuration. Currently, we use a
default configuration which means that the FW decides which antenna to use.

Change-Id: I3134a7e42f264feb6a4405bfe017172789e5afb5
Signed-off-by: Haim Dreyfuss <haim.dreyfuss@intel.com>
iwl7000-tree: 1b88c71685b221a67fa1323cb41921e7d50cbdfb