UPSTREAM: drm/dp_mst: Fill branch->num_ports
This field on drm_dp_mst_branch was never filled

It is initialized to zero when the port is kzallocced.
When a port is added to the list, increment num_ports,
and when a port is removed from the list, decrement num_ports.

v2: remember to decrement on port removal
v3: don't explicitly init to 0
v4: move decrement of num_ports to unlink_port function

Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: David Francis <David.Francis@amd.com>
Signed-off-by: Mikita Lipski <mikita.lipski@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit b1dee9a716b532566570052686144bfbba62aaf6)



BUG=b:170142088
TEST=Compile tested/Backported fixes identified with r-o-m/On-device testing with eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I515ad795c6da8d06e164b702c923011fd18151c3