UPSTREAM: ath10k: Use standard bulk clock API in snoc
No frequency is currently specified for the single clock defined in the
snoc driver, so the clock wrappers reimplements the standard bulk API
provided by the clock framework. Change to this.

The single clock defined is marked as optional so this version of the
get API is used, but might need to be reconsidered in the future.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit f93bcf0ce6a144174292716f9a79330567a778b6)

BUG=b:149055290
TEST=trogdor boots

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: Id67a0454431b9b99c266a80fa8d7d4c62e2d3ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2116116
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>