CHROMIUM: iwl7000: dbg_ini: use vzalloc to allocate dumping memory regions
During dump flow, the driver allocates spaces to store the memory
regions that will be included in the dump. These regions can be very
large so in order to avoid allocation failure, use vzalloc instead.
The kmalloc uses GFP_KERNEL and the driver does not make any use of the
fact that the memory is contiguous so the same functionality is
maintained.

Change-Id: I4b2a001937c2e8cf2bb958e4db887699228cc7c8
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
iwl7000-tree: 539e3cf9c523ddb36836e3ec631fd58fca1efc20