CHROMIUM: iwl7000: pcie: make switch to QnJ config exclusive
Instead of having an inclusive check for QnJ devices, make the
selection exclusive, so that switching to QnJ is the exception, not
the default.  This prevents us from having to add all the non-QnJ
cards to an exclusion list.

To do so, only go into the QnJ block if the device has an RF ID type
HR and HW revision QnJ.

Change-Id: Icc6f8ad2d224690538000255b4946673bb1f60e7
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 68291a1a73e6c3a144a3144f2c5a8183ad45ecbd