UPSTREAM: drm/amd/display: Refactor HDCP to handle multiple displays per link
[Why]
We need to do this to support HDCP over MST

Currently we save a display per link, in a MST case we need to save
multiple displays per link.

[How]
We can create an array per link to cache the displays, but it
complicates the design. Instead we can use the module to cache the
displays.

Now we will always add all the displays to the module, but we use the
adjustment flag to disable hdcp on all of them before they are added.

When we want to enable hdcp we just query the display(cache), remove
it then add it back with different adjustments. Its the similar for
disable.

Signed-off-by: Bhawanpreet Lakha <Bhawanpreet.Lakha@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit b1abe5586ffcb15cc51668bd1a0f97adfbfb0acd)

BUG=b:148404854
TEST=Build and boot to UI

Signed-off-by: Deepak Sharma <Deepak.sharma@amd.com>
Change-Id: I3062a85bcd74b53cf2b1437f4ed5d5bf5b591fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2116682
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Deepak Sharma <deepak.sharma@amd.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Deepak Sharma <deepak.sharma@amd.com>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>