FROMGIT: interconnect: Disallow interconnect core to be built as a module
Building individual drivers as modules is fine but allowing a core
framework to be built as a module makes it really complex and should be
avoided.

Whatever uses the interconnect core APIs must also be built as a module
if interconnect core is built as module, else we will see compilation
failures.

If another core framework (like cpufreq, clk, etc), that can't be built
as module, needs to use interconnect APIs then we will start seeing
compilation failures with allmodconfig configurations as the symbols
(like of_icc_get()) used in other frameworks will not be available in
the built-in image.

Disallow the interconnect core to be built as a module to avoid all
these issues.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/b789cce388dd1f2906492f307dea6780c398bc6a.1567065991.git.viresh.kumar@linaro.org
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
(cherry picked from commit fcb57bfcb87f3bdb1b29fea1a1cd72940fa559fd
 https://git.linaro.org/people/georgi.djakov/linux.git icc-next)

BUG=b:149515210
TEST=generic kernel builders pass

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I96660f036a90a997304d3ed369fdf609ccb2a375
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2233974
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>