UPSTREAM: drm/amd/display: Refactor HDCP encryption status update
[Why]
The old way was to poll PSP and update the properties. But due to a
limitation in the PSP interface this doesn't work for MST.

[How]
According to PSP if set_encryption return success, the link is encrypted
and the only way it will not be is if we get a link loss(which we handle
already).

So this method should be good enough to report HDCP status.

Signed-off-by: Bhawanpreet Lakha <Bhawanpreet.Lakha@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 45375a501f1387567cccd4bbd78d6bbea8d5b68c)

BUG=b:148404854
TEST=Build and boot to UI

Signed-off-by: Deepak Sharma <Deepak.sharma@amd.com>
Change-Id: I14b5bded67ed83ebcef673ffffc071c9cf16c539
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2116684
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Deepak Sharma <deepak.sharma@amd.com>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>