FROMGIT: xhci: remove extra loop in interrupt context
When finishing a TD we walk the endpoint dequeue trb pointer
until it matches the last TRB of the TD.

TDs can contain over 100 TRBs, meaning we call a function 100 times,
do a few comparisons and increase a couple values for each of these calls,
all in interrupt context.

This can all be avoided by adding a pointer to the last TRB segment, and
a number of TRBs in the TD. So instead of walking through each TRB just
set the new dequeue segment, pointer, and number of free TRBs directly.

Getting rid of the while loop also reduces the risk of getting stuck in a
infinite loop in the interrupt handler. Loop relied on valid matching
dequeue and last_trb values to break.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
(cherry picked from commit bf06995629a6f77ca062f53e766aefa948e3dc9a
 git://git.kernel.org/pub/scm/linux/kernel/git/mnyman/xhci.git fuzzing_cleanup)

BUG=b:177098385
TEST=Test read/write on a USB thumb drive

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: Icac99fca1d752682bb27989c2d025d3dda83f483
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2626758
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>