UPSTREAM: drm/amd/display: Change ERROR to WARN for HDCP module
[Why]
HDCP is a bit finicky so we try it 3 times, this leads to a case where
if we fail the first time and pass the second time the error is still
shown in dmesg for the first failed attempt.

This leads to false positive errors.

[How]
Change the logging from ERROR to WARNING. Warnings are still shown in dmesg
to know what went wrong.

Signed-off-by: Bhawanpreet Lakha <Bhawanpreet.Lakha@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 2e956df2913f3452b4d590016f5364fb7255284a)

BUG=b:148404854
TEST=Build and boot to UI

Signed-off-by: Deepak Sharma <Deepak.sharma@amd.com>
Change-Id: I9c44bf2abb867f2cf50d99bbe2c3372404026ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2116679
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Deepak Sharma <deepak.sharma@amd.com>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>